Nest override_type under gen_stub attribute
