+++ examples/pure/src/lib.rs
@@ -252,9 +252,9 @@ fn default_value(num: Number) -> Number {
 
 #[gen_stub_pyfunction]
 #[pyfunction]
-#[override_type(type_repr="collections.abc.Callable[[str]]", imports=("collections.abc"))]
+#[gen_stub(override_type(type_repr="collections.abc.Callable[[str]]", imports=("collections.abc")))]
 fn override_type<'a>(
-    #[override_type(type_repr="collections.abc.Callable[[str]]", imports=("collections.abc"))]
+    #[gen_stub(override_type(type_repr="collections.abc.Callable[[str]]", imports=("collections.abc")))]
     cb: Bound<'a, PyAny>,
 ) -> PyResult<Bound<'a, PyAny>> {
     cb.call1(("Hello!",))?;
