+++ pyo3-stub-gen-derive/src/gen_stub/pymethods.rs
@@ -95,9 +95,7 @@ pub fn prune_attrs(item_impl: &mut ItemImpl) {
             super::attr::prune_attrs(&mut item_fn.attrs);
             for arg in item_fn.sig.inputs.iter_mut() {
                 if let FnArg::Typed(ref mut pat_type) = arg {
-                    pat_type
-                        .attrs
-                        .retain(|attr| !attr.path().is_ident("override_type"));
+                    super::attr::prune_attrs(&mut pat_type.attrs);
                 }
             }
         }
