+++ pyo3-stub-gen-derive/src/gen_stub/util.rs
@@ -3,12 +3,10 @@ use std::collections::HashSet;
 use proc_macro2::TokenStream as TokenStream2;
 use quote::{quote, ToTokens};
 use syn::{
-    parenthesized,
-    parse::{Parse, ParseStream},
-    punctuated::Punctuated,
-    Attribute, GenericArgument, LitStr, PathArguments, PathSegment, Result, ReturnType, Token,
-    Type, TypePath,
+    Attribute, GenericArgument, PathArguments, PathSegment, Result, ReturnType, Type, TypePath,
 };
+
+use crate::gen_stub::attr::parse_gen_stub_override_type;
 
 pub fn quote_option<T: ToTokens>(a: &Option<T>) -> TokenStream2 {
     if let Some(a) = a {
@@ -70,77 +68,14 @@ pub fn extract_return_type(
     };
     let mut ret = ret.clone();
     remove_lifetime(&mut ret);
-    if let Some(r#type) = parse_override_type_attribute(ret.clone(), &attrs)? {
-        return Ok(Some(r#type));
+    if let Some(attr) = parse_gen_stub_override_type(attrs)? {
+        return Ok(Some(TypeOrOverride::OverrideType {
+            r#type: ret.clone(),
+            type_repr: attr.type_repr,
+            imports: attr.imports,
+        }));
     }
     Ok(Some(TypeOrOverride::RustType { r#type: ret }))
-}
-
-/// Parse `override_type` attribute if present.
-pub fn parse_override_type_attribute(
-    r#type: Type,
-    attrs: &Vec<Attribute>,
-) -> Result<Option<TypeOrOverride>> {
-    for attr in attrs {
-        if attr.path().is_ident("override_type") {
-            let attr: OverrideTypeAttribute = attr.parse_args()?;
-            return Ok(Some(TypeOrOverride::OverrideType {
-                r#type,
-                type_repr: attr.type_repr,
-                imports: attr.imports,
-            }));
-        }
-    }
-    Ok(None)
-}
-
-pub struct OverrideTypeAttribute {
-    pub(crate) type_repr: String,
-    pub(crate) imports: HashSet<String>,
-}
-
-mod kw {
-    syn::custom_keyword!(type_repr);
-    syn::custom_keyword!(imports);
-    syn::custom_keyword!(override_type);
-}
-
-impl Parse for OverrideTypeAttribute {
-    fn parse(input: ParseStream) -> Result<Self> {
-        let mut type_repr = None;
-        let mut imports = HashSet::new();
-
-        while !input.is_empty() {
-            let lookahead = input.lookahead1();
-
-            if lookahead.peek(kw::type_repr) {
-                input.parse::<kw::type_repr>()?;
-                input.parse::<Token![=]>()?;
-                type_repr = Some(input.parse::<LitStr>()?);
-            } else if lookahead.peek(kw::imports) {
-                input.parse::<kw::imports>()?;
-                input.parse::<Token![=]>()?;
-
-                let content;
-                parenthesized!(content in input);
-                let parsed_imports = Punctuated::<LitStr, Token![,]>::parse_terminated(&content)?;
-                imports = parsed_imports.into_iter().collect();
-            } else {
-                return Err(lookahead.error());
-            }
-
-            if !input.is_empty() {
-                input.parse::<Token![,]>()?;
-            }
-        }
-
-        Ok(OverrideTypeAttribute {
-            type_repr: type_repr
-                .ok_or_else(|| input.error("missing type_repr"))?
-                .value(),
-            imports: imports.iter().map(|i| i.value()).collect(),
-        })
-    }
 }
 
 fn unwrap_pyresult(ty: &Type) -> &Type {
